move tag blueprint from yafowil.common to yafowil.tag
